Componentizando scaffold messenger e reestruturando páginas de register
