+++ lib/pages/login/login_page.dart
@@ -7,12 +7,11 @@ import 'package:caderneta_campo_digital/controllers/login/login_controller.dart'
 import 'package:caderneta_campo_digital/global/global.dart';
 import 'package:caderneta_campo_digital/pages/home_produtor/home_produtor.dart';
 import 'package:caderneta_campo_digital/pages/home_tecnico/home_tecnico.dart';
-import 'package:caderneta_campo_digital/pages/register/main.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:mask_text_input_formatter/mask_text_input_formatter.dart';
 import 'package:caderneta_campo_digital/pages/select_account/select_account.dart';
-import 'package:caderneta_campo_digital/utils/main.dart';
+import 'package:caderneta_campo_digital/utils/utils.dart';
 import 'package:caderneta_campo_digital/components/UnderlineButton/main.dart';
 
 class LoginPage extends StatefulWidget {
